broke out interface from raxml for generax
